cleanup(core): move lock-file out of utils lib
