+++ packages/nx/src/lock-file/__fixtures__/npm.lock.ts
