+++ packages/nx/src/lock-file/__fixtures__/pnpm.lock.ts
