+++ packages/nx/src/lock-file/__fixtures__/yarn.lock.ts
