+++ packages/nx/src/lock-file/lock-file.spec.ts
