+++ packages/nx/src/lock-file/pnpm.spec.ts
