+++ packages/nx/src/lock-file/pnpm.ts
